room types crud

+++ routes/api.php
@@ -20,48 +20,45 @@ Route::group(['prefix' => 'v1', 'middleware' => 'req.log'], function () {
     });
 
     Route::group(['middleware' => 'api.auth'], function () {
-        Route::group(['prefix' => 'admin', 'middleware' => 'user.is.admin'], function () {
-            //hotel
-            Route::group(['prefix' => 'hotel'], function () {
-                Route::get('', 'HotelController@getHotel');
-                Route::post('image', 'HotelController@addHotelImage');
-                Route::patch('', 'HotelController@updateHotel');
+        Route::group(['prefix' => 'hotel'], function () {
+            Route::get('', 'HotelController@getHotel');
+            Route::post('image', 'HotelController@addHotelImage')->middleware('user.is.admin');
+            Route::patch('', 'HotelController@updateHotel')->middleware('user.is.admin');
+        });
+
+        //rooms
+        Route::group(['prefix' => 'room'], function () {
+            Route::get('', 'RoomController@getRooms');
+            Route::post('', 'RoomController@addRoom')->middleware('user.is.admin');
+
+            Route::group(['middleware' => 'room'], function () {
+                Route::get('/{id}', 'RoomController@getRoom');
+                Route::patch('/{id}', 'RoomController@updateRoom')->middleware('user.is.admin');
+                Route::delete('/{id}', 'RoomController@deleteRoom')->middleware('user.is.admin');
             });
+        });
 
-            //rooms
-            Route::group(['prefix' => 'room'], function () {
-                Route::get('', 'RoomController@getRooms');
-                Route::post('', 'RoomController@addRoom');
+        //room types
+        Route::group(['prefix' => 'room-type'], function () {
+            Route::get('', 'RoomTypeController@getRoomTypes');
+            Route::post('', 'RoomTypeController@addRoomType')->middleware('user.is.admin');
 
-                Route::group(['middleware' => 'room'], function () {
-                    Route::get('/{id}', 'RoomController@getRoom');
-                    Route::patch('/{id}', 'RoomController@updateRoom');
-                    Route::delete('/{id}', 'RoomController@deleteRoom');
-                });
+            Route::group(['middleware' => 'room-type'], function () {
+                Route::get('/{id}', 'RoomTypeController@getRoomType');
+                Route::patch('/{id}', 'RoomTypeController@updateRoomType')->middleware('user.is.admin');
+                Route::delete('/{id}', 'RoomTypeController@deleteRoomType')->middleware('user.is.admin');
             });
+        });
 
-            //room types
-            Route::group(['prefix' => 'room-type'], function () {
-                Route::get('', 'RoomTypeController@getRoomTypes');
-                Route::post('', 'RoomTypeController@addRoomType');
+        //pricing
+        Route::group(['prefix' => 'pricing', 'middleware' => 'user.is.admin'], function () {
+            Route::get('', 'PricingController@getPricings');
+            Route::post('', 'PricingController@addPricing');
 
-                Route::group(['middleware' => 'room-type'], function () {
-                    Route::get('/{id}', 'RoomTypeController@getRoomType');
-                    Route::patch('/{id}', 'RoomTypeController@updateRoomType');
-                    Route::delete('/{id}', 'RoomTypeController@deleteRoomType');
-                });
-            });
-
-            //pricing
-            Route::group(['prefix' => 'pricing'], function () {
-                Route::get('', 'PricingController@getPricings');
-                Route::post('', 'PricingController@addPricing');
-
-                Route::group(['middleware' => 'pricing'], function () {
-                    Route::get('/{id}', 'PricingController@getPricing');
-                    Route::patch('/{id}', 'PricingController@updatePricing');
-                    Route::delete('/{id}', 'PricingController@deletePricing');
-                });
+            Route::group(['middleware' => 'pricing'], function () {
+                Route::get('/{id}', 'PricingController@getPricing');
+                Route::patch('/{id}', 'PricingController@updatePricing');
+                Route::delete('/{id}', 'PricingController@deletePricing');
             });
         });
     });
